requirements.txt
Flask>=3.0,<4.0
google-cloud-firestore>=2.16,<3.0
google-genai>=1.30,<2.0
gunicorn>=22.0,<24.0
openpyxl>=3.1,<4.0
